Seeking and Destroying
Used arguments and filter method
The arguments object includes all the arguments in an input.
This function only has one parameter, yet wants to include more. So the goal is to save all the arguments.
The arguments were transformed into an array, by using array.from.

Use nested for loops to iterate through the block of code. The other for loop is to get the 2nd and 3rd elements of the array.
If elements are equal to either arguments then delete them. Therefore, use an if statement to remove the numbers.
To remove the empty spaces, use the filter method and use an anonymous function. The idea is to see if x is evaluated to true.
Filter iterates through the array, then applies a function to each element of the array. X relates to each element and then filter applies it.
Use boolean to test if true or false.