Convert HTML Entities challenge
Use html entities, .map, string, split and char methods to solve the challenge. The aim is to create objects with key value pairs. Then search for the key and return an argument which is a string. 
NB: Where you have [] then you just return the value. If there isn't a [], then just return the key or the name of the property in the object.
The map function will check through everything in an array to look for a property that matches a key. If there is, then return the value or char key.