Sum All Odd Fibonacci Numbers challenge
Used array.prototype to create an array that was used as an anonymous function. This was used to call on properties to the array objects, last and secondToLast. After this, use a while loop to push values to the end of the array. The filter and reduce methods are used to identify even and odd numbers.