Smallest Common Multiple Challenge
Use the Euclid's algorithm, to find a fast way to get the greatest common divisor (GCD) of two numbers. This means the the largest number that divides both numbers without leaving a remainder. %.
OR solve by using:
A sorted array[] of values. Set the least common multiple - lcm and flag. Use lcm to check the answers.
If yes, then use the flag to return the lcm.