Arguments Optional
Use a function to see if a number is actually a number. If it is fine, if not, return undefined.
Then check the parameters to see if they are numbers. If they are not numbers then use 
a conditional statement - else.
Then return the parameters added together.
If only one parameter was found, return a new function that expects two parameters.
Store first argument before entering the new function.
Check the number again, it must be outside the function to be able to return an object, instead of undefined.
Return function that expects a second argument.
Check for any non-numbers. If none are found, then add numbers.